function pages may not need param ID
If no query param `id` is given, assume we’re only
dealing with projects and return a map without `$id`.
In these cases, `ed` must be passed on (if none is given
in the query, assume we’re dealing with sth. global and
use the data collection’s ID

Cf. 449